now suming up total interactions for a recommended friend with more priority for the direct relation with your friends. Close friends are 1/10 and close friends' friends are 1/15. Trying to take into account each percentage of totalInteractions but the math should be different. For now it's commented out in lines 26, 28 and 30 in Grafo.cpp

+++ Grafo.h
@@ -1,16 +1,14 @@
 #ifndef GRAFO_H_
 #define GRAFO_H_
 
-
 struct Recommendation{
   Vertice *person;
-  int totalInteractions;
-  Recommendation(Vertice *person,int totalInteractions){
+  double totalInteractions;
+  Recommendation(Vertice *person,double totalInteractions){
     this->person = person;
     this->totalInteractions = totalInteractions;
   }
 };
-// deque<Vertice> breadthFirstSearch(Vertice origin,Vertice end);
 vector<Recommendation> recommend(Vertice *origin);
 
 #endif
